fix: Save icons as GIF files (#47)
This change removes drops the reliance on data urls and writes icons to
disk (accessible at ‘/icons/...’). It reduces the main index.html down
to 2MB from ~15MB. It’s till bigger than I’d like but a good start. This
is preparatory work for exporting making the JSON data public as an API.