ci: Use the self-hosted macOS runner (#80)
This change uses the self-hosted macOS runner to improve performance and
allow us to break build encapsulation and take advantage of the runner's
file system to share synced assets between builds. This is a short term
fix to make us more resilient to Internet Archive outages and minimize
the load we put on the original asset sources. In the future, it would
be great to use something a little more formal for this caching and
ultimately switch to a crawler which is robust to outages an cascades
changes through the index.